Added first back button to createBooking. Fixed rating process for homeowner.

+++ application/app/src/main/java/com/project/group/group_project/RatingDatabase.java
@@ -91,7 +91,7 @@ public class RatingDatabase {
             ratingInt += rating.getRate();
         }
 
-        ratingInt = ratingInt / ratings.size();
+        ratingInt = (Integer) (ratingInt / ratings.size());
 
         return ratingInt;
     }
